pyproject.toml
[tool.poetry]
name = "drawitor"
version = "0.1.0"
description = "Draw Images or GIFs in your terminal."
authors = ["Eliaz Bobadilla <eliaz.bobadilladev@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/UltiRequiem/drawitor"
repository = "https://github.com/UltiRequiem/drawitor"
keywords = ["commands", "utilities", "cli"]
include = ["licence"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
drawitor = "drawitor:main"
